refactor: extract auth, rate limiting, and dev endpoints from Program.cs (#111)
AddApplicationAuth(), AddApplicationRateLimiting(), MapDevEndpoints(), and
ConnectionStringBuilder extracted into Infrastructure/. Program.cs reduced to ~85 lines.
